Handle new log formats
Some logs are now defaulting to the 'errorssummary' which doesn't
include actual warnings. We manually replace that with the 'raw' log.
Raw logs are now kind of JSON encoded (the file itself isn't valid) so
we need to parse out just the data portion if possible but still
fallback to a non-JSON path.